intptr_t とかめんどくさいから、long でいいんじゃね的な。どうせintptr_t と long 同じじゃん... (HWND とか HANDLEとかはそのまなの型なんだから)

+++ HmCppInvoke/HmCppInvoke.cpp
@@ -120,7 +120,7 @@ THmMacroVariable Hidemaru::TestDynamicVar = nullptr;
 
 
 // 秀丸の変数が文字列か数値かの判定用
-extern "C" __declspec(dllexport) intptr_t SetDynamicVar(const void* dynamic_value) {
+extern "C" __declspec(dllexport) long SetDynamicVar(const void* dynamic_value) {
 
 	auto param_type = (THm::DLLFUNCPARAM)THm::Hidemaru_GetDllFuncCalledType(1);
 	if (param_type == THm::DLLFUNCPARAM::WCHAR_PTR) {
@@ -128,20 +128,20 @@ extern "C" __declspec(dllexport) intptr_t SetDynamicVar(const void* dynamic_valu
 		return 1;
 	}
 	else {
-		Hidemaru::TestDynamicVar = (intptr_t)dynamic_value;
+		Hidemaru::TestDynamicVar = (long)dynamic_value;
 		return 1;
 	}
 }
 
 
-intptr_t popnumvar = 0;
+long popnumvar = 0;
 // スタックした変数を秀丸マクロから取り出す。内部処理用
-extern "C" __declspec(dllexport) intptr_t PopNumVar() {
+extern "C" __declspec(dllexport) long PopNumVar() {
 	return popnumvar;
 }
 
 // 変数を秀丸マクロから取り出すためにスタック。内部処理用
-extern "C" __declspec(dllexport) intptr_t PushNumVar(intptr_t i_tmp_num) {
+extern "C" __declspec(dllexport) long PushNumVar(long i_tmp_num) {
 	popnumvar = i_tmp_num;
 	return 1;
 }
@@ -153,7 +153,7 @@ extern "C" __declspec(dllexport) const wchar_t* PopStrVar() {
 }
 
 // 変数を秀丸マクロから取り出すためにスタック。内部処理用
-extern "C" __declspec(dllexport) intptr_t PushStrVar(const wchar_t* sz_tmp_str) {
+extern "C" __declspec(dllexport) long PushStrVar(const wchar_t* sz_tmp_str) {
 	popstrvar = sz_tmp_str;
 	return 1;
 }
